Extract CLI code in openapi-codegen-client-fetch
